Moved `matrix_exp` to `heat.core.linalg.basics`
