+++ heat/core/linalg/__init__.py
@@ -10,4 +10,3 @@ from .svd import *
 from .polar import *
 from .eigh import *
 from .randomized import *
-from .other import *
